refactor: create spinner component in sharedComps
